Optimize settings schema generation during release process.
